refactoring: api server
* renamed to admin_server
* split module
