+++ python/src/templates/moves.html
@@ -255,7 +255,13 @@
     </div>
     <div class="row">
         <div class="col">
-            <div style="font-family: monospace; white-space: pre-wrap;">{{apiserver.last_msg}}</div>
+            {% with messages = get_flashed_messages(with_categories=False) %}
+            {% if messages %}
+            {% for message in messages %}
+            {{ message }}
+            {% endfor %}
+            {% endif %}
+            {% endwith %}
         </div>
     </div>
